[866] Use more correct conversion (int/uint32 <-> float) and also fix compile warnings in unix systems
Use more correct conversion (int/uint32 <-> float) and also fix compile warnings in unix systems
Fixed: Murmur check if target has aura from spell before adjusting HP

git-svn-id: https://scriptdev2.svn.sourceforge.net/svnroot/scriptdev2@866 5f9c896b-1e26-0410-94da-f77f675e2462

+++ scripts/zone/hellfire_citadel/shattered_halls/boss_warchief_kargath_bladefist.cpp
@@ -273,12 +273,12 @@ struct MANGOS_DLL_DECL boss_warchief_kargath_bladefistAI : public ScriptedAI
             DoMeleeAttackIfReady();
         }
 
-        if(resetcheck_timer < diff)
+        if (resetcheck_timer < diff)
         {
             uint32 tempx,tempy;
-            tempx = m_creature->GetPositionX();
-            tempy = m_creature->GetPositionY();
-            if ( tempx > 255 || tempx < 205)
+            tempx = uint32(m_creature->GetPositionX());
+            tempy = uint32(m_creature->GetPositionY());
+            if (tempx > 255 || tempx < 205)
             {
                 EnterEvadeMode(); 
             }
